Make partition alignment configurable, bump version to 1.6.0
The L4T R32.4.3 flashing tools force partitions to be aligned
on 4K boundaries (8 sectors), so we'll use that setting by
default to maximize compatibility.

Users that have used previous versions with the old 1-sector
alignment (with L4T R32.5.x, which doesn't impose the 8-sector
alignment in its tooling) can use --with-partition-alignment=1
when configuring.

Signed-off-by: Matt Madison <matt@madison.systems>