Restore the 1-sector alignment default, bump version
Make the default value for --with-partition-alignment 1.
The alignment mismatch mentioned in the last commit was
not due to any tools differences between L4T versions, but
due to how the flash layout XML was being processed by the
recipe that generates the partition table used by this
script. Specifically, the <align_boundary> tag was not
being handled.

That actual root cause has been fixed, so return to
our original default alignment setting.

Update the version to 1.6.1.

Signed-off-by: Matt Madison <matt@madison.systems>