Use 'full_init_payload' for bootloader init, if present
It turns out that NVIDIA's bootloader update tools don't work
properly on t18x systems if the normal BUP payload contain
entries for some of the non-redundant boot partitions, so we've
had to revert the normal bl_update_payload on those platforms
to exclude them.  A separate payload called 'full_init_payload'
will be present in that case, and that should be used instead.

Signed-off-by: Matt Madison <matt@madison.systems>